requirements.txt
annotated-types==0.6.0
attrs==23.2.0
aws-cdk-lib==2.93.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
black==24.1.1
cattrs==23.2.3
click==8.1.7
constructs==10.3.0
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
jsii==1.94.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
publication==0.0.3
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
ruff==0.1.15
six==1.16.0
typeguard==2.13.3
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
